// frontend/src/types/generated/Label.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Label = { id: string, user_id: string | null, name: string, description: string | null, color: string, background_color: string | null, icon: string | null, is_system: boolean, created_at: string, updated_at: string, document_count: number, source_count: number, };